Filter entities by user

+++ src/main/java/eu/marcinszewczyk/web/rest/CommentResource.java
@@ -3,6 +3,7 @@ package eu.marcinszewczyk.web.rest;
 import com.codahale.metrics.annotation.Timed;
 import eu.marcinszewczyk.domain.Comment;
 import eu.marcinszewczyk.repository.CommentRepository;
+import eu.marcinszewczyk.service.UserService;
 import eu.marcinszewczyk.web.rest.errors.BadRequestAlertException;
 import eu.marcinszewczyk.web.rest.util.HeaderUtil;
 import io.github.jhipster.web.util.ResponseUtil;
@@ -16,6 +17,7 @@ import java.net.URISyntaxException;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 /**
  * REST controller for managing Comment.
@@ -29,9 +31,11 @@ public class CommentResource {
     private static final String ENTITY_NAME = "comment";
 
     private final CommentRepository commentRepository;
+    private final UserService userService;
 
-    public CommentResource(CommentRepository commentRepository) {
+    public CommentResource(CommentRepository commentRepository, UserService userService) {
         this.commentRepository = commentRepository;
+        this.userService = userService;
     }
 
     /**
@@ -48,7 +52,7 @@ public class CommentResource {
         if (comment.getId() != null) {
             throw new BadRequestAlertException("A new comment cannot already have an ID", ENTITY_NAME, "idexists");
         }
-        Comment result = commentRepository.save(comment);
+        Comment result = commentRepository.save(comment.user(userService.getCurrentUserId()));
         return ResponseEntity.created(new URI("/api/comments/" + result.getId()))
             .headers(HeaderUtil.createEntityCreationAlert(ENTITY_NAME, result.getId().toString()))
             .body(result);
@@ -70,6 +74,9 @@ public class CommentResource {
         if (comment.getId() == null) {
             throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
         }
+        if (!userService.verifyEntityUser(comment)) {
+            throw new BadRequestAlertException("It's not yours!", ENTITY_NAME, "user");
+        }
         Comment result = commentRepository.save(comment);
         return ResponseEntity.ok()
             .headers(HeaderUtil.createEntityUpdateAlert(ENTITY_NAME, comment.getId().toString()))
@@ -85,7 +92,9 @@ public class CommentResource {
     @Timed
     public List<Comment> getAllComments() {
         log.debug("REST request to get all Comments");
-        return commentRepository.findAll();
+        return commentRepository.findAll().stream()
+            .filter(userService::verifyEntityUser)
+            .collect(Collectors.toList());
     }
 
     /**
@@ -99,6 +108,9 @@ public class CommentResource {
     public ResponseEntity<Comment> getComment(@PathVariable Long id) {
         log.debug("REST request to get Comment : {}", id);
         Optional<Comment> comment = commentRepository.findById(id);
+        if (comment.isPresent() && !userService.verifyEntityUser(comment.get())) {
+            throw new BadRequestAlertException("It's not yours!", ENTITY_NAME, "user");
+        }
         return ResponseUtil.wrapOrNotFound(comment);
     }
 
@@ -112,6 +124,11 @@ public class CommentResource {
     @Timed
     public ResponseEntity<Void> deleteComment(@PathVariable Long id) {
         log.debug("REST request to delete Comment : {}", id);
+
+        Optional<Comment> task = commentRepository.findById(id);
+        if (task.isPresent() && !userService.verifyEntityUser(task.get())) {
+            throw new BadRequestAlertException("It's not yours!", ENTITY_NAME, "user");
+        }
 
         commentRepository.deleteById(id);
         return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert(ENTITY_NAME, id.toString())).build();
